+++ lib/src/controls/inputs/buttons/filled_button.dart
@@ -8,7 +8,7 @@ import 'package:fluent_ui/fluent_ui.dart';
 ///
 ///   * [Button], the default button
 ///   * [OutlinedButton], an outlined button
-///   * [TextButton], a borderless button with mainly text-based content
+///   * [HyperlinkButton], a borderless button with mainly text-based content
 class FilledButton extends Button {
   /// Creates a filled button
   const FilledButton({